backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
langchain==0.3.0
langchain-anthropic==0.2.4
anthropic==0.40.0
dj-database-url==2.1.0
google-auth==2.28.0
djangorestframework-simplejwt==5.3.1
